Added the drif_corr2 code to the main code (fracture_tracking). Removed drif_corr2.m
drift correction on basis of binning the displacement of particles and subsequently selecting a value for drift was not in the project overview. Furthermore, this displacement-binning method is a lot more complex than just taking average positions, so there are multiple sections of code for the drift correction at the moment (see section description): one for drift correction based on binning the displacement, one for using average displacement (didn't look good), and one for the  average position subtracted from the original position (not finished).